Bring LOCKMON into AUTO and create MOMON.
MOMON turns off all motors if there isn't a thread executing and
momon=1 (to allow for raw commands to work). This is in it's own thread
incase it ever issues the command when the motors arent stopped. If it
does it will casuea command error and auto will restart the monitor.